@@ -1,57 +1,25 @@
-import React from 'react';
+import React, { useEffect, useState } from "react";
 
-class Clock extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      clockId: 0,
-      time: 0
-    };
-  }
+function Clock({ started, tick }) {
+  const [value, setValue] = useState(0);
 
-  startClock() {
-    var clockId = setInterval(this.clockTick.bind(this), 1000);
-    this.setState({ clockId: clockId });
-  }
+  useEffect(() => {
+    let interval = null;
 
-  pauseClock() {
-    clearInterval(this.state.clockId);
-    this.setState({ clockId: null });
-  }
-
-  resetClock() {
-    clearInterval(this.state.clockId);
-    this.setState({ time: 0, clockId: null });
-  }
-
-  clockTick() {
-    this.setState({ time: ++this.state.time });
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if (this.props.status === nextProps.status)
-      return;
-
-    switch (nextProps.status) {
-      case 'new':
-        this.resetClock();
-        break;
-      case 'playing':
-        this.startClock();
-        break;
-      case 'winner':
-      case 'loser':
-      default:
-        this.pauseClock();
-        break;
+    if (started) {
+      interval = setInterval(() => {
+        setValue(tick());
+      }, 1000);
     }
-  }
 
-  render() {
-    return (
-      <div className="clock">Time: {this.state.time}</div>
-    );
-  }
+    return () => clearInterval(interval);
+  }, [started, tick]);
+
+  return (
+    <div className="clock">
+      Time: {value}
+    </div>
+  );
 }
 
 export default Clock;
