@@ -1,12 +1,17 @@
-import React from 'react';
-import Button from './Button';
-import '../styles/Difficulty.css';
+import React from "react";
+import "./Difficulty.css";
+import { useGame } from "./GameContext";
 
-const Difficulty = ({value}) =>
-  <div className="difficulty">
-    <Button value={0} active={value === 0} text="Beginner" />
-    <Button value={1} active={value === 1} text="Intermediate" />
-    <Button value={2} active={value === 2} text="Expert" />
-  </div>;
+function Difficulty() {
+  const { newGame } = useGame();
+
+  return (
+    <div className="difficulty">
+      <button onClick={() => newGame(0)}>Beginner</button>
+      <button onClick={() => newGame(1)}>Intermediate</button>
+      <button onClick={() => newGame(2)}>Expert</button>
+    </div>
+  );
+}
 
 export default Difficulty;
