@@ -1,13 +1,17 @@
 import React from 'react';
-import {render} from 'react-dom';
-import {Provider} from 'react-redux';
+import ReactDOM from 'react-dom';
 import App from './components/App';
-import store from './stores';
-import registerServiceWorker from './registerServiceWorker';
+import reportWebVitals from './reportWebVitals';
+import "./index.css";
 
-render(
-  <Provider store={store}><App /></Provider>,
-  document.getElementById('app')
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
 );
 
-registerServiceWorker();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
